Adds setTotalSteps() method to ProgressBar.
The idea behind is is to set stepSize to a value that will evenly divide this progress bard into specified amount of steps.

Math.Abs() is called after division just in case totalSteps is negative.

I'm not sure if totalSteps should be a float (might be useful for someone) or int, but I decided that int does have the most sense.